requirements.txt
psycopg2
pandas
requests
matplotlib
jinja2
pyyaml
langchain
llama-cpp-python
